Add login page and switch to task view on login
