+++ static/src/main.ts
@@ -12,11 +12,7 @@ import Toaster from "@/plugins/toaster";
 OpenAPI.BASE =
   import.meta.env.VITE_API_URL?.toString().replace(/\/$/, "") || "";
 
-const app = createApp(App, {
-  taskId: "ecg-qrs-classification-physiodb",
-  competency: "Prof. Dr. Med.",
-  user: "AnnotoUser#1337",
-});
+const app = createApp(App);
 
 app.use(Inkline, {
   components,
